Completely reworked the Na'vi data structure. Now both OfflineNavi.kt and OnlineNavi.kt will get converted into the universal Navi.kt structure which will be passed to the UI. This new universal structure is much more performant and cleaner to handle and view. Complicated Reykunyu syntax is parsed into custom data classes, enums and lists.
For offline dictionary, the conversion will happen only if the word is searched, improving performance, this also offloading heavy tasks such as Regex from the UI to the search coroutine, improving runtime performance.
the online Na'vi (v2) structure is mostly completed. Just need to add the downloading logic and write a custom deserializer for the pesky etymology json.

All done in a day's work :)